backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "alex-database",
    "boto3>=1.40.29",
    "google-cloud-aiplatform>=1.148.0",
    "langfuse>=3.3.4",
    "openai-agents>=0.3.0",
    "pydantic-ai>=1.0.6",
    "python-dotenv>=1.1.1",
]

[tool.uv.workspace]
members = [
    "database",
    "api",
    "scheduler",
]

[tool.uv.sources]
alex-database = { workspace = true }
